[refactor] allow accessing the consensus commit monitor independently. Extract the authority building method and reuse across containers.
