+++ consensus/core/src/swarm/container.rs
@@ -1,17 +1,11 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 use super::node::NodeConfig;
+use crate::{CommitConsumerMonitor, CommittedSubDag, ConsensusAuthority, TransactionClient};
 use arc_swap::ArcSwapOption;
-use consensus_config::Parameters;
 use mysten_metrics::monitored_mpsc::UnboundedReceiver;
-use prometheus::Registry;
-use std::{sync::Arc, time::Duration};
+use std::sync::Arc;
 use tracing::info;
-
-use crate::transaction::NoopTransactionVerifier;
-use crate::{
-    CommitConsumer, CommitConsumerMonitor, CommittedSubDag, ConsensusAuthority, TransactionClient,
-};
 
 pub(crate) struct AuthorityNodeContainer {
     consensus_authority: ConsensusAuthority,
@@ -24,7 +18,7 @@ impl AuthorityNodeContainer {
     pub async fn spawn(config: NodeConfig) -> Self {
         info!(index =% config.authority_index, "starting in-memory node non-sim");
         let (consensus_authority, commit_receiver, commit_consumer_monitor) =
-            Self::make_authority(config).await;
+            super::node::make_authority(config).await;
 
         Self {
             consensus_authority,
@@ -33,16 +27,7 @@ impl AuthorityNodeContainer {
         }
     }
 
-    pub fn take_commit_receiver(
-        &self,
-    ) -> (
-        UnboundedReceiver<CommittedSubDag>,
-        Arc<CommitConsumerMonitor>,
-    ) {
-        let commit_consumer_monitor = self
-            .commit_consumer_monitor
-            .swap(None)
-            .expect("Commit consumer has been already consumed");
+    pub fn take_commit_receiver(&self) -> UnboundedReceiver<CommittedSubDag> {
         let commit_receiver = self
             .commit_receiver
             .swap(None)
@@ -50,7 +35,14 @@ impl AuthorityNodeContainer {
         let Ok(commit_receiver) = Arc::try_unwrap(commit_receiver) else {
             panic!("Commit receiver already consumed");
         };
-        (commit_receiver, commit_consumer_monitor)
+        commit_receiver
+    }
+
+    pub fn commit_consumer_monitor(&self) -> Arc<CommitConsumerMonitor> {
+        self.commit_consumer_monitor
+            .load_full()
+            .expect("Commit consumer monitor should be available")
+            .clone()
     }
 
     pub fn transaction_client(&self) -> Arc<TransactionClient> {
@@ -59,59 +51,5 @@ impl AuthorityNodeContainer {
 
     pub fn is_alive(&self) -> bool {
         true
-    }
-
-    async fn make_authority(
-        config: NodeConfig,
-    ) -> (
-        ConsensusAuthority,
-        UnboundedReceiver<CommittedSubDag>,
-        Arc<CommitConsumerMonitor>,
-    ) {
-        let NodeConfig {
-            authority_index,
-            db_dir,
-            committee,
-            keypairs,
-            network_type,
-            boot_counter,
-            protocol_config,
-        } = config;
-
-        let registry = Registry::new();
-
-        // Cache less blocks to exercise commit sync.
-        let parameters = Parameters {
-            db_path: db_dir.path().to_path_buf(),
-            dag_state_cached_rounds: 5,
-            commit_sync_parallel_fetches: 2,
-            commit_sync_batch_size: 3,
-            sync_last_known_own_block_timeout: Duration::from_millis(2_000),
-            ..Default::default()
-        };
-        let txn_verifier = NoopTransactionVerifier {};
-
-        let protocol_keypair = keypairs[authority_index].1.clone();
-        let network_keypair = keypairs[authority_index].0.clone();
-
-        let (commit_consumer, commit_receiver, _) = CommitConsumer::new(0);
-        let commit_consumer_monitor = commit_consumer.monitor();
-
-        let authority = ConsensusAuthority::start(
-            network_type,
-            authority_index,
-            committee,
-            parameters,
-            protocol_config,
-            protocol_keypair,
-            network_keypair,
-            Arc::new(txn_verifier),
-            commit_consumer,
-            registry,
-            boot_counter,
-        )
-        .await;
-
-        (authority, commit_receiver, commit_consumer_monitor)
     }
 }
